+++ do_text_header.php
@@ -265,12 +265,9 @@ function save_audio_position($textid): void
         const text_id = parseInt(<?php echo json_encode($textid); ?>, 10);
         $.ajax({
             type: "POST",
-            url:'api.php/v1/texts/' + text_id + '/reading-position',
+            url:'api.php/v1/texts/' + text_id + '/audio-position',
             data: {
-                action: "reading_position",
-                action_type: "audio",
-                tid: text_id,
-                audio_position: pos
+                position: pos
             }, 
             async: false // Asynchronous should be safe (2.9.0)
         });